backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
outcome==1.3.0.post0
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
